fix: exit method on cancelled permission request
A permission request is cancelled when there is another one pending.
In the case that happens, the onRequestPermissionResult method is called with permissions and gratResults arrays empty, which should make the method to return immediately, avoiding the getLocation method to be called with no permissions granted.

References https://outsystemsrd.atlassian.net/browse/RNMT-4280