fix: ask user to change location settings when necessary
Instead of returning "Current location settings can not satisfy this request" , show dialog so that the user can enable the necessary location settings.

Still missing onActivityResult to handle the user's response in the dialog, and, if location settings are changed, try to call the getLocation.

Note: Also updated import for NonNull